ocr_pipeline/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
python-multipart>=0.0.9
pydantic-settings>=2.0
pdf2image>=1.17.0
Pillow>=10.0
PyMuPDF>=1.24.0
tiktoken>=0.7.0
websockets>=12.0
httpx>=0.27.0
